Shopping Cart API, mailing list working

+++ Workspace/javaBootcamp2017/src/main/java/Clase4Services/IndividualItem.java
@@ -4,7 +4,7 @@ public class IndividualItem implements Item {
 	private String name;
 	private double price;
 	
-	public IndividualItem(String aName, int aPrice){
+	public IndividualItem(String aName, double aPrice){
 		this.name= aName;
 		this.price= aPrice;
 	}
